+++ app/src/main/java/com/example/moviesearch/FavoritesFragment.kt
@@ -1,16 +1,15 @@
 package com.example.moviesearch
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
+import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.moviesearch.databinding.FragmentFavoritesBinding
 
 
-class FavoritesFragment : Fragment() {
+class FavoritesFragment (val filmsDataBase: List<Film>): Fragment() {
     private lateinit var binding: FragmentFavoritesBinding
     private lateinit var filmsAdapter: FilmListRecyclerAdapter
     override fun onCreateView(
@@ -23,7 +22,7 @@ class FavoritesFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val favoritesList: List<Film> = emptyList()
+        val favoritesList: List<Film> = filmsDataBase.filter { it.isInFavorites }
 
         binding.favoritesRecycler.apply {
             filmsAdapter =
@@ -38,13 +37,8 @@ class FavoritesFragment : Fragment() {
             addItemDecoration(decorator)
         }
         filmsAdapter.addItems(favoritesList)
-        updateData(newList = ArrayList(favoritesList))
+
     }
 
-    private fun updateData(newList: ArrayList<Film>) {
-        val filmDiff = FilmDiff(oldList = ArrayList(newList), newList)
-        val diffResult = DiffUtil.calculateDiff(filmDiff)
-        filmsAdapter.addItems(newList)
-        diffResult.dispatchUpdatesTo(filmsAdapter)
-    }
+
 }
