+++ app/src/main/res/values/strings.xml
@@ -1,14 +1,12 @@
 <resources>
     <string name="app_name">World Movies</string>
 
-    <string name="source_of_posters">Posters taken from the site: https://yandex.ru/images/</string>
     <string name="favorites">Favorites</string>
     <string name="watch_later">Watch Later</string>
     <string name="selections">Selections</string>
     <string name="settings">Settings</string>
     <string name="find_your_movie">Find Your Movie</string>
     <string name="toast_settings">Settings</string>
-    <string name="toast_favorites">Favorites</string>
     <string name="toast_watchlater">Watch later</string>
     <string name="toast_selections">Selections</string>
     <string name="description">Description</string>
@@ -18,6 +16,5 @@
     <string name="alertdialog_negative">No</string>
     <string name="snack_favorites">Added to \"Favorites\"</string>
     <string name="snack_favorites_deleted">Deleted from \"Favorites\"</string>
-    <string name="snack_watch_later">Added to \"Watch Later\"</string>
 
 </resources>
